+++ src/irvine/oeis/a003/A003052.java
@@ -20,7 +20,7 @@ public class A003052 extends Sequence1 {
   public static boolean isColumbian(final long n) {
     boolean ok = true;
     for (int k = 0; k <= n; ++k) {
-      if (k + Functions.DIGIT_SUM.l((long) k) == n) {
+      if (k + Functions.DIGIT_SUM.l(k) == n) {
         ok = false;
         break;
       }
