+++ src/irvine/oeis/a007/A007436.java
@@ -18,7 +18,7 @@ public class A007436 extends Sequence1 {
     Z sum = Z.ZERO;
     for (final Z dd : Jaguar.factor(++mN).divisors()) {
       final int d = dd.intValue();
-      sum = sum.add(Functions.FIBONACCI.z((long) d).multiply(Functions.MOBIUS.i((long) (mN / d))));
+      sum = sum.add(Functions.FIBONACCI.z(d).multiply(Functions.MOBIUS.i((long) (mN / d))));
     }
     return sum;
   }
