+++ src/irvine/oeis/a046/A046209.java
@@ -15,6 +15,6 @@ public class A046209 extends Sequence1 {
 
   @Override
   public Z next() {
-    return Integers.SINGLETON.sumdiv(++mN, d -> Z.THREE.pow(mN / d).multiply(Functions.GCD.i(d, 3)).multiply(Functions.MOBIUS.i((long) d))).divide(3L * mN);
+    return Integers.SINGLETON.sumdiv(++mN, d -> Z.THREE.pow(mN / d).multiply(Functions.GCD.i(d, 3)).multiply(Functions.MOBIUS.i(d))).divide(3L * mN);
   }
 }
